requirements.txt
PyQt5
numpy
pycryptodome
argon2-cffi
qrcode
Pillow
pyzbar
opencv-python
